feat: add header component
