+++ packages/landingpage-new/components/layout.module.scss
@@ -1,16 +1,3 @@
 .appContainer {
-  margin: 32px 45px;
-}
-
-.header {
-  height: 120px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.logoContainer {
-  display: flex;
-  align-items: center;
-  gap: 8px;
+  margin: 24px 45px;
 }
